Fix default thread not going to game thread when not explicitly chosen Make dequeue_max_count uint32_t

+++ cppmods/EventViewerMod/include/Structs.h
@@ -116,7 +116,7 @@ namespace RC::EventViewerMod
         EMiddlewareHookTarget hook_target = EMiddlewareHookTarget::All;          // [Savable] [Thread-ImGui]
         EMode mode = EMode::Stack;                                               // [Savable] [Thread-ImGui]
         uint16_t dequeue_max_ms = 10;                                            // [Savable] [Thread-ImGui]
-        uint16_t dequeue_max_count = 50000;                                      // [Savable] [Thread-ImGui]
+        uint32_t dequeue_max_count = 50000;                                      // [Savable] [Thread-ImGui]
 
         std::string blacklist;                                                   // [Savable] [Thread-ImGui]
         std::vector<std::string> blacklist_tokens;                               // [Thread-ImGui] (lower-cased tokens)
